Split Benchee integration tests
